// src/server/types.gen.ts
export type ModelsServer = {
    admin_id?: string;
    created_at?: string;
    current_players?: number;
    description?: string;
    host?: string;
    id?: string;
    max_players?: number;
    name?: string;
    port?: number;
    status?: string;
    updated_at?: string;
    version?: string;
};